requirements.txt
easydict==1.9 
imageio==2.19.3 
matplotlib==3.5.2
matplotlib-inline 
mkl-fft==1.3.1 
opencv-python==4.6.0.66  
Pillow==9.0.1 
PyYAML==5.4.1 
scikit-learn==1.1.1
scipy==1.8.1 
tensorboard==1.14.0
tensorboardX==2.5.1 
torch==1.4.0
torchvision==0.5.0 
tqdm==4.64.0 